requirements.txt
anthropic
streamlit
chromadb
sentence-transformers
requests
beautifulsoup4
pymupdf
python-docx
python-dotenv